requirements_hf.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
statsmodels>=0.14.0
requests>=2.31.0
pyyaml>=6.0
joblib>=1.3.0
Pillow>=10.0.0
gradio>=4.0.0
